scripts/requirements.txt
backoff==1.10.0
gitpython==3.1.9
lxml==4.5.2
pathspec==0.8.0
phabricator==0.7.0
pyaml==20.4.0
requests==2.24.0
retrying==1.3.3
unidiff==0.6.0
python-benedict==0.22.0